requirements.txt
pytesseract
opencv-python
Pillow
pdf2image
